fix(suite): circular deps in connect popup
