+++ scripts/circular-dependencies/madge-snapshot.txt
@@ -73,10 +73,6 @@ packages/suite/src/utils/wallet/graph/utilsShared.ts > packages/suite/src/utils/
 packages/suite/src/views/dashboard/AssetsView/AssetTable/AssetTable.tsx > packages/suite/src/views/dashboard/AssetsView/AssetsView.tsx
 packages/suite/src/views/onboarding/steps/SelectBackupType/SelectBackupType.tsx > packages/suite/src/views/onboarding/steps/SelectBackupType/FloatingSelections.tsx
 packages/transport/src/transports/abstract.ts > packages/transport/src/transports/bridge.ts
-suite-common/connect-popup/src/methodHooks/index.ts > suite-common/connect-popup/src/methodHooks/addressConfirmation.ts
-suite-common/connect-popup/src/methodHooks/index.ts > suite-common/connect-popup/src/methodHooks/bitcoinSignTransaction.ts
-suite-common/connect-popup/src/methodHooks/index.ts > suite-common/connect-popup/src/methodHooks/ethereumSignTransaction.ts
-suite-common/connect-popup/src/methodHooks/index.ts > suite-common/connect-popup/src/methodHooks/solanaSignTransaction.ts
 suite-common/graph/src/graphDataFetching.ts > suite-common/graph/src/graphUtils.ts
 suite-common/intl-types/src/index.ts > suite-common/intl-types/src/types.ts > packages/suite/src/components/suite/Translation.tsx
 suite-common/trading/src/reducers/tradingReducer.ts > suite-common/trading/src/thunks/buy/index.ts > suite-common/trading/src/thunks/buy/confirmBuyTradeThunk.ts
