+++ suite-common/connect-popup/src/methodHooks/index.ts
@@ -1,33 +1,11 @@
-import { Dispatch } from '@reduxjs/toolkit/react';
-
-import { PrecomposedTransactionFinal } from '@suite-common/wallet-types';
-import {
-    CallMethodKeys,
-    CallMethodParams,
-    CallMethodResponse,
-    Success,
-    Unsuccessful,
-} from '@trezor/connect';
+import { CallMethodKeys } from '@trezor/connect';
 
 import { addressConfirmationModalHooks } from './addressConfirmation';
 import { bitcoinSignTransaction } from './bitcoinSignTransaction';
 import { ethereumSignTransaction } from './ethereumSignTransaction';
 import { requestLoginHooks } from './requestLogin';
 import { solanaSignTransaction } from './solanaSignTransaction';
-import { ConnectCallSource } from '../connectPopupTypes';
-
-export type PreCallHookParams<M extends CallMethodKeys> = {
-    method: M;
-    payload: Omit<CallMethodParams<M>, 'method'>;
-    dispatch: Dispatch;
-    getState: () => any;
-    txSigningPrecomposed?: PrecomposedTransactionFinal;
-    source: ConnectCallSource;
-};
-export type PostCallHookParams<M extends CallMethodKeys> = PreCallHookParams<M> & {
-    originalPayload: Omit<CallMethodParams<M>, 'method'>;
-    response: Success<CallMethodResponse<M>> | Unsuccessful;
-};
+import { PostCallHookParams, PreCallHookParams } from './types';
 
 export const preCallHooks = async <M extends CallMethodKeys>(params: PreCallHookParams<M>) => {
     await bitcoinSignTransaction.preCallHook(params);
