+++ suite-common/connect-popup/src/methodHooks/requestLogin.ts
@@ -1,6 +1,6 @@
 import type { CallMethodKeys } from '@trezor/connect';
 
-import { PreCallHookParams } from './index';
+import { PreCallHookParams } from './types';
 
 const preCallHook = <M extends CallMethodKeys>({
     method,
